Refactor Device Metric component
